+++ src/components/content/conditions/helpers/history.tsx
@@ -1,7 +1,6 @@
 import { SearchParams } from "fhir-kit-client";
 import { HistoryEntryProps } from "../../resource/helpers/history-entry";
 import { useHistory } from "../../resource/history";
-import { CodingList } from "@/components/core/coding-list";
 import { NotesList } from "@/components/core/notes-list";
 import { ConditionModel } from "@/fhir/models";
 import { capitalize, startCase } from "@/utils/nodash";
@@ -68,20 +67,22 @@ function getHistoryEntry(condition: ConditionModel): HistoryEntryProps {
       value: capitalize(condition.verificationStatus),
     },
     {
-      label: "Recorded Date",
-      value: condition.recordedDate,
-    },
-    {
       label: "Category",
-      value: startCase(condition.categories[0]),
+      value: startCase(condition.category),
     },
     {
       label: "Note",
       value: condition.notes.length !== 0 && <NotesList notes={condition.notes} />,
     },
     {
-      label: "Code",
-      value: <CodingList codings={condition.knownCodings} />,
+      label: "Recorded Code(s)",
+      value: (
+        <div className="ctw-space-y-1">
+          {condition.knownCodings.map((coding) => (
+            <div key={`${coding.system}-${coding.code}`}>{coding.display}</div>
+          ))}
+        </div>
+      ),
     },
     {
       label: "Onset Date",
@@ -101,8 +102,7 @@ function getHistoryEntry(condition: ConditionModel): HistoryEntryProps {
     id: condition.id,
     date: condition.recordedDate,
     versionId: condition.versionId,
-    title: startCase(condition.categories[0]),
-    subtitle: condition.patientOrganizationName,
+    title: condition.patientOrganizationName,
     details: detailData,
   };
 }
@@ -113,7 +113,7 @@ const valuesToDedupeOn = (condition: ConditionModel) => [
   condition.verificationStatus,
   condition.recordedDate,
   condition.notes,
-  condition.categories[0],
+  condition.category,
   condition.onset,
   condition.abatement,
   condition.encounter,